+++ python/moveit_py/src/moveit_py/moveit_ros/moveit_cpp/planning_component.cpp
@@ -34,7 +34,8 @@
 
 /* Author: Peter David Fagan */
 
-#include <utils.h>
+#include <copy_ros_msg.h>
+#include <serialize_ros_msg.h>
 #include "planning_component.h"
