+++ python/moveit_py/src/moveit_py/utils/copy_ros_msg.cpp
@@ -35,7 +35,7 @@
 /* Author: Peter David Fagan */
 
 #include <list>
-#include "utils.h"
+#include "copy_ros_msg.h"
 
 // Ros Message Copy Definitions (Note: copying faster than serialize/deserialize)
 
@@ -394,4 +394,4 @@ moveit_msgs::msg::Constraints ConstraintsToCpp(py::object constraints)
 //  }
 
 //  return rosmsg;
-//}
+//}_
